+++ ManaKit/Classes/CardTableViewCell.swift
@@ -82,7 +82,7 @@ public class CardTableViewCell: UITableViewCell {
         }
         
         // name
-        nameLabel.text = ManaKit.sharedInstance.name(ofCard: card)
+        nameLabel.text = card.displayName
         
         if let releaseDate = card.set!.releaseDate {
             let isModern = ManaKit.sharedInstance.isModern(card)